move article.ts -> Article.ts
